pyproject.toml
[tool.poetry]
name = "fastapi-postgres-sqlalchemy"
version = "0.1.0"
description = ""
authors = ["kwiats <pawuskwiatkowski@gmail.com>"]
readme = "README.md"
packages = [{include = "fastapi_postgres_sqlalchemy"}]

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "^0.18.3"
fastapi = "^0.83.0"
pydantic = "^1.10.2"
SQLAlchemy = "^1.4.41"
psycopg2-binary = "^2.9.3"
SQLAlchemy-Utils = "^0.38.3"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
